apply review fixes (staleTime, type narrow, util move, comment)
