+++ src/config/constants.ts
@@ -81,4 +81,4 @@ export const CARD_COLUMNS = 3;
 export const CARD_ROWS = 3;
 export const CARD_LIMIT = CARD_COLUMNS * CARD_ROWS;
 
-export const MAX_PAGE_LIMIT = 120;
+export const MAX_PAGE_LIMIT = 120; // BE hard cap per page
