+++ src/utils/copyEmails.ts
